requirements.txt
python-dotenv
gradio
langchain_text_splitters
langchain_openai
langchain_huggingface
langchain_chroma
langchain_community